refactor(dispatcher): split dispatcher.py into a directory module
Closes #128.

The single 1169-line `dispatcher.py` file grew five distinct concerns
since #66 (job model, UI dispatcher, standalone dispatcher, idle pump,
cooperative-cancellation probe).  This commit reorganises them into a
proper directory module so each piece can evolve in isolation and so
upstreaming candidates (e.g. `BaseDccCallableDispatcher` in
dcc-mcp-core 0.14.17 #521) have a clean cut-over surface.

Layout
------
::

    src/dcc_mcp_maya/dispatcher/
    ├── __init__.py        # zero-overhead public re-exports
    ├── job.py             # _JobEntry + _current_job ContextVar
    ├── cancel.py          # check_maya_cancelled probe
    ├── ui.py              # MayaUiDispatcher (interactive)
    ├── standalone.py      # MayaStandaloneDispatcher (mayapy / batch)
    └── pump.py            # MayaUiPump + _CorePump + factories

Public API surface preserved 1:1 — every name previously importable
from `dcc_mcp_maya.dispatcher` still resolves to the same object.
`__all__` is declared explicitly in `__init__.py` so the public
contract is now machine-readable.

Verification
------------
* All 54 dispatcher tests pass unchanged.
* No external import statements needed updating: callers already write
  `from dcc_mcp_maya.dispatcher import X` and the package re-export
  preserves that contract.
* Each new submodule is independently importable and ruff-clean.
* The previous file is removed (no dual-source ambiguity).
